rearrange files move class decor and react decor to root level
